* added initial super admin addresses to genesis file

* super holds only 1 token

* added test info store

* OLP-1082: Health Passport: Fork Protocol Code
-First commit

* OLP-1082: Health Passport: Fork Protocol Code
-Fixed VT token transfer issue, re-enabled send tx and currency list.

* OLP-1082: Health Passport: Fork Protocol Code
-Fixed decimal issue with olclient sendfunds

* initial implementation of uploadTx

* initialize auth token store

* initialize auth token store

* added remove token interface

* updated token store key format

* added unit test for auth token store

* clean up old code

* OLP-1085: Create Hospital Admin Auth Token. First commit.

* updated upload Tx

* added some update

* data layer of ReadTx

* OLP-1085: Create Hospital Admin Auth Token.

* OLP-1085: Create Hospital Admin Auth Token.
Added error codes and super user account data for testing.

* OLP-1085: Create Hospital Admin Auth Token.
Closing wallet for secure signing service

* implemented data layer for read Tx

* updated iteration interface

* resolved conflicts coming from develop branch

* corrected typo

* updated Tx request structs

* PR view update

* OLP-1086: Auth Token for Screener Admin Tx

* basic setup

* OLP-1086: Auth Token for Screener Admin Tx
- Removed chainstate reads from validate functions

* upload test info Tx

* basic setup

* basic setup

* changes to create_hospital_admin.go

* Launching Loadtest through shell script

* deelted unsused shell script

* modified shell script with additional config

* modified shell script with additional config

* modified shell script with additional config

* resolved conflicts with develop branch

* modified shell script with additional config

* fixing token permissions for screener.

* modified shell script with additional config

* renamed virusType as testType

* makefile update

* code review update

* modified shell script with additional config

* modified shell script with additional config

* test commit

* merged latest

* read test Tx implementation

* added validation for super user.

* added extra validation for create hospital and screener admin tokens.

* Added validation for timestamp

* Added validation for timestamp

* refactored loadtest framework

* Added function to validate timestamp

* integration tests on read data Tx

* merged develop branch

* refactored loadtest framework

* refactored loadtest framework

* load tests for uploadTx and readTx

* added error messages for parallel reads

* add Cleanup() back which was deleted accidentally

* Let screener admin sign scanTx instead of the person

* code review

* change and add some field names, also put some as mandatory

* update TestInfo struct

* synchronized with Hao's branch

* added UpdateTestInfo() interface to store

* reordered fields in the structs

* done test, before merging charlie's branch

* uses SH256 hashes for super admin IDs

* integration tests for update test info result

* Olp 1226 Create Tx change Mandatory Fields (#14)

* change and add some field names, also put some as mandatory

* update TestInfo struct

* synchronized with Hao's branch

* added UpdateTestInfo() interface to store

* reordered fields in the structs

* done test, before merging charlie's branch

* add query filter for analysisOrg, analyzedBy and analyzed time

* clean code


* updated load tests: update test info & super admins SHA256 userID

* cleaned up printing

* Hot fix (#16)

* fix null pointer

* fix for read.go


* Cleanup (#17)

* cleanup

* fix json tag


* add permission to query: get auth tokens by org (#18)


* remove err to solve the nil pointer problem (#19)


* Olp 1283 Separate update test info logic to a new tx (#20)

* separate update logic to a different tx

* integration test

* integration test done

* remove person address check

* monor fix

* change address check to optional

* remove test org from update tx


* remove the limit for same person with multiple auth token (#21)


* Olp 1394 IHP: Create a Separate Call for user to fetch their data with UUID (#22)

* create a query for user to query results without permission check

* test done

+++ data/keys/key_store.go
@@ -7,8 +7,8 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"encoding/json"
-	"errors"
 	"fmt"
+	"github.com/pkg/errors"
 	"io"
 	"io/ioutil"
 	"os"
@@ -132,6 +132,12 @@ func (ks *KeyStore) decrypt(data []byte, passphrase string) ([]byte, error) {
 }
 
 func (ks *KeyStore) SaveKeyData(path string, address Address, data []byte, passphrase string) error {
+	//Check if keyfile already exists
+	existingFile, _ := GetFileName(path, address)
+	if len(existingFile) > 0 {
+		return errors.New("key file already exists")
+	}
+
 	filename, _ := filepath.Abs(path + buildFileName(address))
 	f, _ := os.Create(filename)
 
@@ -172,10 +178,9 @@ func (ks *KeyStore) GetKeyData(path string, address Address, passphrase string) 
 	if err != nil {
 		return nil, err
 	}
-
 	err = json.Unmarshal(data, accData)
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "Parallel reads not permitted on keystore file")
 	}
 
 	cipherData, err := hex.DecodeString(accData.CipherText)
@@ -283,7 +288,7 @@ func GetFileName(path string, address Address) (string, error) {
 	}
 
 	if len(matches) != 1 {
-		return "", errors.New("invalid address")
+		return "", errors.New("invalid address - path: " + path + " address: " + address.String())
 	}
 
 	return matches[0], nil
